[buteo-syncfw] Make static the creation of profile from XML
When using the SyncClientInterface, several
signals and functions pass profiles as XML
strings. Creating a profile on the fly from
these strings already exist in the
ProfileManager class, and does not depend on
any member of the class. Make it static to
ease its usage in conjunction with a
SyncClientInterface object.